Implement mod loading for Linux native and Proton games
Linux native game launching uses an additional "stage 2" wrapper that is
injected into the game's sandbox after all of Steam's wrappers. It is
then able to launch the game binary with the doorstop native
library injected.

The Tauri build seems to break if we put the workspace Cargo.toml in the
root of the project, rather than in the `crates` directory. Not ideal,
but it works.